base: move here the script for autocomplete
* Moves the script for autocompletion from groups module to
  base module.

* Modifies the autocompletion script to automatically enable
  the functionality after the page loading.
