+++ invenio/modules/groups/bundles.py
@@ -24,8 +24,7 @@ from invenio.base.bundles import jquery as _j, invenio as _i
 
 js = Bundle(
     'vendors/jquery/dist/jquery.js',
-    'vendors/typeahead.js/dist/typeahead.bundle.js',
-    'js/groups/init.js',
+    'js/remote.autocomplete.field.js',
     filters=RequireJSFilter(exclude=[_j, _i]),
     output="groups.js",
     weight=50
